Avoid passing templates/files over RPC
Pre-store the template for nested stacks in the database and use the new
template_id parameter in the the RPC API to pass the ID over RPC,
instead of passing a copy of the whole template and all the files. This
will help us to safely de-duplicate templates in the database, reduce
the size of messages, and prevent unnecessary copies of the files (in
particular) being loaded in memory.

Change-Id: I1305572cd70845a28df60756440dd417733b0da1