Add root_id column to stack table
Before this patch, calls to Stack.get_root_id() ended up performing a
recursive db operation with as many calls as the depth of the (nested)
stack. This is mainly invoked when resources are created; if there
are 100 resources at an average stack depth of 20, that's about 2000
more db calls than we ought to need. With this patch, none of those
calls are necessary.

This patch also helps with quickly counting total resources
from anywhere within the nested stack.

Change-Id: I09b79e96900544e8633e7052ea21799502d18d4d