Don't load all rsrc defns for every check_resource
In the case where we are initially loading a resource in convergence
(check_resource in worker.py), it is not necessary to load any other
of the stack's resource definitions.

We are still creating all the stack's resource definitions (as we need
to) whenever construct_input_data() or _updated_stored_properties()
is called.

Change-Id: Ib54f3118ec6801184db28f07333e25b886b1521d